Changed bridge import and made all queues 1 to avoid delays
