+++ projects/opendr_ws_2/src/opendr_perception/opendr_perception/face_detection_retinaface_node.py
@@ -22,52 +22,53 @@ import numpy as np
 
 from sensor_msgs.msg import Image as ROS_Image
 from vision_msgs.msg import Detection2DArray
-from ros2_bridge.bridge import ROS2Bridge
+from opendr_ros2_bridge import ROS2Bridge
 
+from opendr.perception.object_detection_2d import RetinaFaceLearner
+from opendr.perception.object_detection_2d import draw_bounding_boxes
 from opendr.engine.data import Image
-from opendr.perception.object_detection_2d import DetrLearner
-from opendr.perception.object_detection_2d.detr.algorithm.util.draw import draw
 
 
-class ObjectDetectionDetrNode(Node):
+class FaceDetectionNode(Node):
 
     def __init__(self, input_image_topic="image_raw", output_image_topic="/opendr/image_boxes_annotated",
-                 detections_topic="/opendr/objects", device="cuda"):
-        super().__init__('object_detection_detr_node')
+                 face_detections_topic="/opendr/faces", device="cuda", backbone="resnet"):
+        super().__init__('face_detection_node')
 
         if output_image_topic is not None:
-            self.image_publisher = self.create_publisher(ROS_Image, output_image_topic, 10)
+            self.image_publisher = self.create_publisher(ROS_Image, output_image_topic, 1)
         else:
             self.image_publisher = None
 
-        if detections_topic is not None:
-            self.bbox_publisher = self.create_publisher(Detection2DArray, detections_topic, 10)
+        if face_detections_topic is not None:
+            self.face_publisher = self.create_publisher(Detection2DArray, face_detections_topic, 1)
         else:
-            self.bbox_publisher = None
-
-        self.image_subscriber = self.create_subscription(ROS_Image, input_image_topic, self.callback, 10)
+            self.face_publisher = None
+        self.image_subscriber = self.create_subscription(ROS_Image, input_image_topic, self.callback, 1)
 
         self.bridge = ROS2Bridge()
 
-        self.detr_learner = DetrLearner(device=device)
-        self.detr_learner.download(path=".", verbose=True)
+        self.face_detector = RetinaFaceLearner(backbone=backbone, device=device)
+        self.face_detector.download(path=".", verbose=True)
+        self.face_detector.load("retinaface_{}".format(backbone))
+        self.class_names = ["face", "masked_face"]
 
     def callback(self, data):
         image = self.bridge.from_ros_image(data, encoding='bgr8')
         cv2.imshow("image", image.opencv())
-        cv2.waitKey(5)
+        cv2.waitKey(1)
 
-        boxes = self.detr_learner.infer(image)
+        boxes = self.face_detector.infer(image)
 
-        image = np.float32(image.opencv())
+        image = image.opencv()
 
-        # Convert detected boxes to ROS type and publish
-        ros_boxes = self.bridge.to_ros_bounding_box_list(boxes)
-        if self.bbox_publisher is not None:
-            self.bbox_publisher.publish(ros_boxes)
+        ros_boxes = self.bridge.to_ros_boxes(boxes)
+        if self.face_publisher is not None:
+            self.face_publisher.publish(ros_boxes)
 
         # Annotate image and publish result
-        image = draw(image, boxes)
+        odr_boxes = self.bridge.from_ros_boxes(ros_boxes)
+        image = draw_bounding_boxes(image, odr_boxes, class_names=self.class_names)
         if self.image_publisher is not None:
             self.image_publisher.publish(self.bridge.to_ros_image(Image(image), encoding='bgr8'))
 
@@ -84,14 +85,14 @@ def main(args=None):
     except:
         device = 'cpu'
 
-    object_detection_detr_node = ObjectDetectionDetrNode(device=device)
+    face_detection_node = FaceDetectionNode(device=device, backbone="resnet")
 
-    rclpy.spin(object_detection_detr_node)
+    rclpy.spin(face_detection_node)
 
     # Destroy the node explicitly
     # (optional - otherwise it will be done automatically
     # when the garbage collector destroys the node object)
-    object_detection_detr_node.destroy_node()
+    face_detection_node.destroy_node()
     rclpy.shutdown()
